streamlit_energy_app/requirements.txt
streamlit 
pandas 
matplotlib 
numpy 
scikit-learn 
scipy
seaborn 
plotly
statsmodels
python-dotenv
sqlalchemy 
pymongo
passlib[bcrypt]
torch
torchvision
prophet
cmdstanpy
joblib
itsdangerous



